# docs/swagger.yaml
definitions:
  stability_ai.BalanceResponse:
    properties:
      credits:
        type: number
    type: object
  stability_ai.BaseErrorResponse:
    properties:
      message:
        type: string
    type: object
  stability_ai.GenerateInput:
    properties:
      cfg_scale:
        default: 7
        description: 확산 프로세스가 프롬프트 텍스트에 얼마나 엄격하게 적용되는지 (값이 높을수록 이미지가 프롬프트에 더 가깝게 유지됨)
        maximum: 35
        minimum: 0
        type: integer
      clip_guidance_preset:
        enum:
        - FAST_BLUE
        - FAST_GREEN
        - NONE
        - SIMPLE
        - SLOW
        - SLOWER
        - SLOWEST
        type: string
      height:
        default: 512
        maximum: 2048
        minimum: 512
        type: integer
      samples:
        default: 1
        description: 생성할 이미지 수
        maximum: 10
        minimum: 1
        type: integer
      steps:
        default: 40
        description: 실행할 확산 단계 수. 이미지 품질에 영향 큼. 비용에도 영향 큼.
        maximum: 150
        minimum: 10
        type: integer
      style:
        description: 개발 중. 생략해도 됨
        type: string
      text_prompts:
        type: string
      width:
        default: 512
        maximum: 2048
        minimum: 512
        type: integer
    required:
    - cfg_scale
    - clip_guidance_preset
    - height
    - samples
    - steps
    - text_prompts
    - width
    type: object
  stability_ai.Image:
    properties:
      imageID:
        description: 이미지 고유 ID
        type: string
      keywords:
        description: 이미지의 키워드 (프롬프트에서 추출)
        items:
          type: string
        type: array
      requestInfo:
        $ref: '#/definitions/stability_ai.StabilityApiPayload'
        description: 이미지 생성 요청값
      time:
        type: string
      url:
        description: 이미지 URL
        type: string
    type: object
  stability_ai.StabilityApiPayload:
    properties:
      cfg_scale:
        type: integer
      clip_guidance_preset:
        type: string
      height:
        type: integer
      samples:
        type: integer
      steps:
        type: integer
      text_prompts:
        items:
          $ref: '#/definitions/stability_ai.TextPrompt'
        type: array
      width:
        type: integer
    required:
    - cfg_scale
    - samples
    - text_prompts
    type: object
  stability_ai.TextPrompt:
    properties:
      text:
        type: string
      weight:
        default: 1
        type: integer
    type: object
info:
  contact: {}
paths:
  /v1/balance:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/stability_ai.BalanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/stability_ai.BaseErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/stability_ai.BaseErrorResponse'
      summary: stability.ai 잔여 크레딧 조회
      tags:
      - credit
  /v1/generate/{engineId}:
    post:
      consumes:
      - application/json
      parameters:
      - description: 엔진 ID
        enum:
        - stable-diffusion-512-v2-1
        - stable-diffusion-512-v2-0
        - stable-diffusion-v1-5
        in: path
        name: engineId
        required: true
        type: string
      - description: 상세는 아래 Model 클릭해서 정보 확인
        in: body
        name: GenerateInput
        required: true
        schema:
          $ref: '#/definitions/stability_ai.GenerateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/stability_ai.Image'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/stability_ai.BaseErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/stability_ai.BaseErrorResponse'
      summary: 이미지 생성하기
      tags:
      - image
  /v1/image:
    get:
      consumes:
      - application/json
      parameters:
      - description: (Optional) keyword로 조건 검색
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/stability_ai.Image'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/stability_ai.BaseErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/stability_ai.BaseErrorResponse'
      summary: 이미지 리스트 조회
      tags:
      - image
  /v1/image/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: image의 uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/stability_ai.Image'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/stability_ai.BaseErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/stability_ai.BaseErrorResponse'
      summary: ID로 이미지 조회
      tags:
      - image
swagger: "2.0"